trying to z buffer

+++ src/Vertex.java
@@ -53,8 +53,7 @@ public class Vertex {
         z += dz;
     }
 
-
-    public void rotateXAxis(double theta, int rY, int rZ) {
+    public void rotateXAxis(double theta, double rY, double rZ) {
         translate(0, -rY, -rZ);
         double thetaR = toRadians(theta);
         double yP = y * cos(thetaR) - z * sin(thetaR);
@@ -64,7 +63,7 @@ public class Vertex {
         translate(0, rY, rZ);
     }
 
-    public void rotateYAxis(double theta, int rX, int rZ) {
+    public void rotateYAxis(double theta, double rX, double rZ) {
         translate(-rX, 0, -rZ);
         double thetaR = toRadians(theta);
         double xP = x * cos(thetaR) + z * sin(thetaR);
@@ -74,7 +73,7 @@ public class Vertex {
         translate(rX, 0, rZ);
     }
 
-    public void rotateZAxis(double theta, int rX, int rY) {
+    public void rotateZAxis(double theta, double rX, double rY) {
         translate(-rX, -rY, 0);
         double thetaR = toRadians(theta);
         double xP = x * cos(thetaR) - y * sin(thetaR);
